fixed small bug in backpropagate

+++ MCTS.hpp
@@ -10,7 +10,7 @@ private:
    std::shared_ptr<Node> bestChild(std::shared_ptr<Node> node, int currentPlayer);
    std::shared_ptr<Node> expand(std::shared_ptr<Node> node);
    double simulate(std::shared_ptr<Node> root);
-   void backpropagate(std::shared_ptr<Node> end, int player, int result);
+   void backpropagate(std::shared_ptr<Node> end, int player, double result);
 
 public:
    MCTS(std::shared_ptr<Game> game);
